hal: fix no sound issue in voice call
No audio is observed in voice call scenario since in_call state
is not update when voice usecase is stopped. This results in not
starting voice call usecase when setting up next call.

Fix by setting in_call state as false in voice_stop_usecase(). This
is to ensure that voice usecase is stopped and in_call state is update
properly.

Conflicts:
	hal/voice.c

CRs-Fixed: 996510
Change-Id: I480bb14899d678a47df8a61fbdd208b75ef7b395