audio: Ignored connect/disconnect events for remote submix devices
When remote submix input is started, audio policy switches the device to
AUDIO_DEVICE_OUT_REMOTE_SUBMIX and sends connect event to HAL modules.
set_parameter call for connect sometimes waits for longer duration to
acquire lock if start_output_stream for a compress stream is in progress
at the same time. This is causing failure of testRecordRemoteSubmix
GTS test case

Fix is to ignore connect/disconnect events for remote submix devices
in Primary Audio HAL as these devices are not applicable

Change-Id: I98f8632c9267b40a09fe83c5816f79e59563f937