hal: standby offload stream on receiving cmd error
OffloadThread invalidates tracks when it receives error notification
from async callback thread which will close current output and
reopen a new one.

If there are concurrent usecases active on the same device when the
error occurs it is possible some state maintained in the HAL won't
get reset(e.g previous sound device won't get disabled and reenabled)

Fix this by putting stream on standby and in effect disabling devices
as soon as cmd error is received instead of waiting for closeOutput
from AudioTrack to standby the stream.

CRs-Fixed: 2131661
Change-Id: I7f9d59a0a4f70f048b06553252da3ba15d7e9d79