hal: fix position query issue for track offload
Currently the real time when position queried is returned as
framesAt for track offload, while the actual framesAt should
be earlier when query happens within the write cycle, this can
lead drift in calculated played duration. Thus A/V sync is
mismatched.
Fix this by returning with the rendered time instead of the
current real time when position is queried.

Change-Id: I0837c7db8ec996dde50fd91c50052090b46bdbf1
CRs-Fixed: 2167391