configs: disable vehicle focus prop by default
There have been multiple issues wrt no audio
from codec, which are identified related to the
enforcement of audio focus validation in vendor
audio HAL.
Recent clarification with Andorid Automotive PoC
suggests that focus should not be enforced and
it's not a required contract.
Thus setting the default to false to disable audio
focus prop.

Change-Id: I102f3529b457083573f19ff55b15d1fd50f5bcc8