hal: Fix to play ring/alarm during 44.1k native stream playback
ringtone/alarm were not heard on speaker and heard only on
headphone during 44.1 native stream playback on headphone due
to invalid device id's were used to check if native playback
is supported on that device and if that device is a headset
combo device. Hence invalid sampling rate was set to speaker
backend and ringtone/alarm were not heard on it.

Added change to set speaker backend sampling rate to 48k to
play ringtone/alarm on speaker.

CRs-Fixed: 2373202
Change-Id: I91aa8dde613ee1a01dd138d0dfa193bd7a3c2350