HAL: Asynchronous PARTIAL_DRAIN implementation
Implementing asynchronous PARTIAL_DRAIN thread for PCM in Gapless Playback

As PCM is decoded data, it does not take much time for the DSP to process it
due to which thei early EOS callback event is sent back at almost the end of stream.

This implementation created a thread that sends the callback event while
sending data down to be rendered.

Signed-off-by: Viraja Kommaraju <virajak@codeaurora.org>

Change-Id: I8082dbfb115085246f40713f83fb7b0010a4419c