HAL: fix external modules search path
In LE, Yocto will pass ${libdir} to configure.ac automatically. And when
multilib feature is disabled, the value of ${libdir} is constant /usr/lib
irrespective of machine architecture, i.e., library installation path
depends on multilib availability.
So far, audiohal searches externl libs in either /usr/lib or /usr/lib64,
depending on the MACRO __LP64__. But in the situation when multilib
is disabled and machine is 64-bit, audiohal can't load external modules
properly because the search path is hard-coded to /usr/lib64 while the
libs are actually installed in /usr/lib.
Thus, set the lib search path dynamically according to ${libdir} for LE,
so that the code can work in different situations.

Change-Id: I054db5c13df88a183db7c671d422f9420b79e96b